Fix usage line in help page to be path to program
This commit adjusts the help screen to display the path to this program rather
than the name. Normally, this is the same as the binary, but in some cases it
can be different.

Conventionally, the path is shown. For instance, GNU cat does the following:

    $ cat --help
    Usage: cat [OPTION]... [FILE]...

Which is not the name of the program, but the path:

    $ ln -s /usr/bin/cat dog
    $ ./dog --help
    Usage: ./dog [OPTION]... [FILE]...
    Concatenate FILE(s) to standard output.